Fix #86 - contract.py does way too much, refactor it
