+++ pdr_backend/util/contract.py
@@ -1,0 +1,96 @@
+import artifacts
+import glob
+import hashlib
+import json
+import os
+from pathlib import Path
+import time
+
+import addresses
+from eth_account import Account
+from eth_account.signers.local import LocalAccount
+from eth_keys import KeyAPI
+from eth_keys.backends import NativeECCBackend
+from sapphirepy import wrapper
+from web3 import Web3, HTTPProvider, WebsocketProvider
+from web3.logs import DISCARD
+from web3.middleware import construct_sign_and_send_raw_middleware
+
+from pdr_backend.util.constants import (
+    ZERO_ADDRESS,
+    SAPPHIRE_TESTNET_CHAINID,
+    SAPPHIRE_MAINNET_CHAINID,
+    MAX_UINT,
+)
+
+keys = KeyAPI(NativeECCBackend)
+
+
+def get_address(chain_id, contract_name):
+    network = get_addresses(chain_id)
+    if not network:
+        raise ValueError(f"Cannot figure out {contract_name} address")
+    address = network.get(contract_name)
+    return address
+
+
+def get_addresses(chain_id):
+    address_filename = os.getenv("ADDRESS_FILE")
+    path = None
+    if address_filename:
+        address_filename = os.path.expanduser(address_filename)
+        path = Path(address_filename)
+    else:
+        path = Path(str(os.path.dirname(addresses.__file__)) + "/address.json")
+
+    if not path.exists():
+        raise TypeError(f"Cannot find address.json file at {path}")
+
+    with open(path) as f:
+        data = json.load(f)
+    for name in data:
+        network = data[name]
+        if network["chainId"] == chain_id:
+            return network
+    return None
+
+
+def get_contract_abi(contract_name):
+    """Returns the abi for a contract name."""
+    path = get_contract_filename(contract_name)
+
+    if not path.exists():
+        raise TypeError("Contract name does not exist in artifacts.")
+
+    with open(path) as f:
+        data = json.load(f)
+        return data["abi"]
+
+
+def get_contract_filename(contract_name):
+    """Returns abi for a contract."""
+    contract_basename = f"{contract_name}.json"
+
+    # first, try to find locally
+    address_filename = os.getenv("ADDRESS_FILE")
+    path = None
+    if address_filename:
+        address_filename = os.path.expanduser(address_filename)
+        address_dir = os.path.dirname(address_filename)
+        root_dir = os.path.join(address_dir, "..")
+        paths = Path(root_dir).rglob(contract_basename)
+        paths = [str(path) for path in paths]
+        if paths:
+            assert len(paths) == 1, "had duplicates for {contract_basename}"
+            path = paths[0]
+            path = Path(path).expanduser().resolve()
+            assert (
+                path.exists()
+            ), f"Found path = '{path}' via glob, yet path.exists() is False"
+            return path
+    # didn't find locally, so use use artifacts lib
+    path = os.path.join(os.path.dirname(artifacts.__file__), "", contract_basename)
+    path = Path(path).expanduser().resolve()
+    if not path.exists():
+        raise TypeError(f"Contract '{contract_name}' not found in artifacts.")
+    return path
