+++ pdr_backend/util/test/test_subgraph.py
@@ -4,7 +4,7 @@ from enforce_typing import enforce_types
 import pytest
 from web3 import Web3
 
-from pdr_backend.utils.subgraph import (
+from pdr_backend.util.subgraph import (
     key_to_725,
     value_to_725,
     value_from_725,
